Split `IRemote` to `IBarRemote` (#3182)
