+++ src/Maestro/Maestro.DataProviders/MaestroBarClient.cs
@@ -13,7 +13,6 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
 
 namespace Maestro.DataProviders;
