+++ src/Microsoft.DotNet.Darc/Darc/Operations/DefaultChannelStatusOperation.cs
@@ -35,7 +35,7 @@ class DefaultChannelStatusOperation : UpdateDefaultChannelBaseOperation
             return Constants.ErrorCode;
         }
 
-        IRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
+        IBarRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
 
         try
         {
