+++ src/Microsoft.DotNet.Darc/Darc/Operations/DeleteChannelOperation.cs
@@ -30,7 +30,7 @@ internal class DeleteChannelOperation : Operation
     {
         try
         {
-            IRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
+            IBarRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
 
             // Get the ID of the channel with the specified name.
             Channel existingChannel = (await remote.GetChannelsAsync()).Where(channel => channel.Name.Equals(_options.Name, StringComparison.OrdinalIgnoreCase)).FirstOrDefault();
