+++ src/Microsoft.DotNet.Darc/Darc/Operations/GetChannelsOperation.cs
@@ -33,7 +33,7 @@ internal class GetChannelsOperation : Operation
     {
         try
         {
-            IRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
+            IBarRemote remote = RemoteFactory.GetBarOnlyRemote(_options, Logger);
 
             var allChannels = await remote.GetChannelsAsync();
             switch (_options.OutputFormat)
