+++ src/Microsoft.DotNet.Darc/Darc/Options/SubscriptionsCommandLineOptions.cs
@@ -55,7 +55,7 @@ abstract class SubscriptionsCommandLineOptions : CommandLineOptions
     [Option("ids", Separator = ',', HelpText = "Get only subscriptions with these ids.")]
     public IEnumerable<string> SubscriptionIds { get; set; }
 
-    public async Task<IEnumerable<Subscription>> FilterSubscriptions(IRemote remote)
+    public async Task<IEnumerable<Subscription>> FilterSubscriptions(IBarRemote remote)
     {
         IEnumerable<DefaultChannel> defaultChannels = await remote.GetDefaultChannelsAsync();
         return (await remote.GetSubscriptionsAsync()).Where(subscription =>
